Add GUI elements to display and edit game object scale
Added code to display and edit the scale of the selected game object in the GUI. Introduced a new section in the `render` method to show the scale using `ImGui::Text`. Retrieved the current scale of the selected game object using `GetScale` and stored it in a `glm::vec3` variable. Created a `float` array to hold the scale values for use with `ImGui::InputFloat3`. Added an `ImGui::InputFloat3` widget to allow users to input new scale values. Commented out the line that would set the new scale values to the selected game object, similar to the existing commented-out rotation setting.